重构App.vue，合并Header和Footer组件内容，删除冗余组件，分离上传进度估算功能
