--- src/components/Footer.vue
@@ -1,18 +1,0 @@
-<template>
-    <footer>
-        <p>
-            © CJackHwang.保留所有权利.
-        </p>
-        <p>
-            <a href="https://github.com/CJackHwang/Fuck-Codemao-Detection" target="_blank">项目仓库：FlowChunkFlex</a>
-        </p>
-        <p>
-            <a href="https://github.com/CJackHwang">
-                <img src="https://img.shields.io/badge/GitHub-CJackHwang-100000?style=flat&logo=github&logoColor=white" alt="GitHub Badge">
-            </a>
-            <a href="https://www.gnu.org/licenses/gpl-3.0.html">
-                <img src="https://img.shields.io/badge/License-GPL%203.0-blue.svg?style=flat" alt="GPL-3.0 License Badge">
-            </a>
-        </p>
-    </footer>
-</template>
