--- src/components/Header.vue
@@ -1,9 +1,0 @@
-<template>
-    <header>
-        <h2>FCF 微云盘</h2>
-        <a>Ver：5.1.5</a>
-        <p>
-            基于Vue3+Vite开发；上传时流式读取文件，分块并发提交至CodemaoCDN以绕过大小限制；中断自动重传，最大单块大小20MB，最大并发上传数3。
-        </p>
-    </header>
-</template>
